Use helper for successful responses
Summary:
# Context
All successful requests have a fixed structure defined by the protocol specification.

# Problem
We are contructing a successful response manually everywhere; which means that new contributors would need to read the spec to check they are building it correctly or blindly copy and paste.

# This diff
Let's just add constructors with a clear name. This also makes the call sites arguably a bit easier to read

Reviewed By: TheGeorge

Differential Revision: D70486084

fbshipit-source-id: aeb437ebbeb7d9b2e3b536ac7d93c3c6d6bdba34