Eqwalize edb_dap_test_support module
Summary:
# Context
The `edb_dap_test_support` module contains many utility functions we use when writing tests for the DAP server

# Problem
Because it is part of the `test` directory, it doesn't get eqwalized by default, and had already many wrong specs. It is faster to refactor things if we get eqwalizer signal than failed tests.

# This diff
Opt in the module to eqwalizer, and fix all the issues.

Reviewed By: TheGeorge

Differential Revision: D70491149

fbshipit-source-id: a7739a9aa30d0ae79be3c410220d32f675b9b1ff