Step-over and step-out should fail if they can never succeed
Summary: Stepping out of a function requires setting breakpoints in the caller. Before this patch, this could silently failed if (for example) the caller is part of OTP and was not compiled with beam_debug_info. This would lead to execution just resuming and not stopping afterwards. With this patch, the step-out command cleanly fails instead. Similarly, by putting a breakpoint in one process, we can pause another process, which is executing code without beam_debug_info. With this patch, trying to step-over in such code fails cleanly.

Reviewed By: jcpetruzza

Differential Revision: D70089605

fbshipit-source-id: 3af9075bd5a7f0e76324a726e31ecfb39aca9411