Improve handling of debuggee modules
Summary:
# Context
In many of edb suites, we start a debuggee node using `peer`, make some custom module available on it, and start a process that runs that module, typically to hit a breakpoints.

To do this, we rely on a function that, given a peer node, and some specification of a module (by contents or by a file), compiles it and loads the module.

# Problem
There are currently two main problems with this approach:

1. Because we are always ensuring the module is loaded, we don't really test the fact that we are supposed to ensure the module is loaded before setting a breakpoint; and
2. We want to add a "configuration phase" where the debuggee is paused while the client sends all the initial breakpoints. But if we then try to load a module this way while the debuggee is paused, the test blocks since the process backing `peer` are also paused.

# This diff
So we then change the way module handling as follows:

1. Every time we start a peer, we create a dedicated private directory for the peer, where we will include a `src` and a `ebin` directory for modules for that peer. The `ebin` directory is added to the codepath, so it is enough to compile a `src` file and save it there, for the module to be available (without forcing it to be loaded)
 2. We also extend the functions to create peers to receive a list of custom modules for the peer. That way, we get a fully initialized node, and we don't need to do further calls to make modules available.

We then rewrite the necessary tests so that they use the new interface.

Reviewed By: robertoaloi

Differential Revision: D70565638

fbshipit-source-id: b2312c977a5332bea47ea06fd57518b106b22366