Make convenience function for constructing common errors
Summary:
# Context
In request handlers, we usually have to return errors due to "user error".

# Problem
For some common classes, we are still constructing the "reaction" term manually, which requires including `edb_dap.hrl` to get the right error code and, is ultimately, too low-level

# This diff
We add a couple more convenience functions to `edb_dap_request` to streamline the process of returning errors.

Reviewed By: robertoaloi

Differential Revision: D70619876

fbshipit-source-id: 8859527bfc6f3c98b9cd1d77524364d42b6472e5