Handle configuration phase of the protocol
Summary:
# Context
The DAP specification says that, after the debugger sends an `initialized` event to indicate that

> [it] is ready to accept configuration requests (e.g. `setBreakpoints`, `setExceptionBreakpoints`).

How the configuration phase is ended depends on whether both client and server  support the `configurationDone` request. When this is not the case, the client sends a final `setExceptionBreakpoints` request even if no filter has been defined by the server, to indicate it is done configuring.

For details, see the docs for the [Initialized event](https://microsoft.github.io/debug-adapter-protocol/specification#Events_Initialized)

# Problem
At the moment, we don't support the `configurationDone` request, but also we don't pay attention to the configuration phase. In particular, after attaching to the debuggee, we let it run while the client sends us the initial `setBreakpoints` requests. This means there is a race-condition, in which, by the time we receive the `setBreakpoints` request, the debuggee has already executed that point. The fact that we can trust breakpoints for the "launch" case is pretty bad.

# This diff
As a first step, we add basic support for the configuration phase. What we do:

  - After attaching to a node (either via "attach" or "launch"), we pause the node and enter the `configuring` state.
  - We remain in the `configuring` state until we receive a `setExceptionBreakpoints` request (necessarily empty, as we don't handle them yet). At that point, we resume the node with all the breakpoints correctly configured.

Notice that at this point:

  1. We don't yet honor the configurationDone message; this is now simple to do, but will happen on a subsequent diff.
  2. There is still a race-condition when "launching" the debuggee node, as the execution could have gone past the breakpoint before we manage to attach. This will be definitely solved by moving to reverse-attach.

Reviewed By: robertoaloi

Differential Revision: D70565459

fbshipit-source-id: ce3642461cf6c5f086e3b198c6d776e27c71f0e6