Support the configurationDone request
Summary:
# Context
In D70565459 we added support for the configuration-phase. In that version, we used the first `setExceptionBreakpoints` as a marker that the configuration phase is over, which corresponds to the case when the server doesn't report that it supports the "configurationDoneRequest" capability.

# Problem
The DAP specification says that instead the client can send a `configurationDone` request when it's done sending all breakpoint-related messages, etc. This is a more reasonable protocol than relying on the first `setExceptionBreakpoints` message, but the server needs to support it.

# This diff
So we add support for the `configurationDone` request and move the logic from the handler for `setExceptionBreakpoints` to the new handler.

I initially expected that the client would have to support the `configurationDone` capability as well, so we'd have to consider that case in the `setExceptionBreakpoints` handler, but there is no such field sent in the [`initialize` request](https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Initialize). We can somehow assume the client is always new enough, it seems.

Reviewed By: robertoaloi

Differential Revision: D70623751

fbshipit-source-id: 4f5acfcac8065f0c00f3e15f55ec8c7e29075578