Fix incorrect handling of stripSourcePrefix in attach request
Summary:
# Context
In the "attach" and "launch" requests, we can pass a `stripSourcePrefix` option to control how we build an absolute path from the beam source, whenever this one is relative.

# Problem
When we introduced the "attach" request handler we somehow used `stripSourcePath` in some places, instead of `stripSourcePrefix`. The way the code is written, we were likely to fail the calls to `edb:attach()` whenever `stripSourcePrefix` was actually given.

Why wasn't this caught by a test? We don't currently have tests for the stripSourcePrefix behaviour. It is a bit annoying to handle as it depends on the beam files containing a relative source source path, which happens with buck2 and otherwise needs quite a bit of faff to get (`+deterministic` and `+{compile_info, [{source, ....}]}`).

# This diff
We fix the issue, and in the process a fix the spec for arguments. We don't add a test for this, for now, not to overcomplicate the setup

Reviewed By: robertoaloi

Differential Revision: D70628016

fbshipit-source-id: 67f850c63daa52a7ac8ebafbc1b5b781dd6d9cd6