Add unit tests for argument parser lib
Summary:
# Context
The `edb_dap_parse` module is an utility module to support parsing of the JSON requests we get from the client.

This started as a very ad-hoc way of validating the "launch" request arguments (as that is not part of the DAP spec), and evolved to be a small parser-combinator library we use to also "parse" other requests (which includes converting fields from binary to atom, etc).

# Problem
Because it started as an ad-hoc thing for the "launch" request, we never added a suite for it, and instead have in edb_dap_launch_config_SUITE tests for the parsing of that particular config, were we are testing the same thing multiple times, just to cover many of the paths of the parsing library. This doesn't scale very well.

# This diff
We add a test suite for edb_dap_parser. In the process we identify a couple of corner cases we were not handling correctly and add additional combinators that we will use next.

Reviewed By: alanz

Differential Revision: D70704465

fbshipit-source-id: 7447b83d0970e3dccb7365fb081489c899600168