Make debuggee initialization code injection transparent by default
Summary:
# Context
When launching a new debuggee node, we need to inject a small code snippet to the node so that it performs a "reverse-attach" to edb.

At the moment, the "launch" request passes this code in an env var to the caller in the `runInTerminal` reverse-request. This let's the caller handle complex scenarios like when the debuggee node runs on a separate node (e.g. buck2 test-shell) or some initialization needs to be done before the codepath is set (rebar3 shell)

# Problem
Whlie this handles the the complex case, it also means that the user needs to deal with this also in the simple cases, while it would be nice that in those cases, things happen automagically

# This diff
We make it so by default, an `-eval CODE_TO_INJECT` is added to the `ERL_AFLAGS` environment variable, and provide a configuration to ask that instead it is put in an environment variable, that the caller names.

Notice that it is important to use `ERL_AFLAGS` instead of `ERL_FLAGS`, as otherwise, in a case like launching:

```
erl -eval foo:bar() -s erlang halt
````

if we were using `ERL_FLAGS` our code would be injected after `erlang:halt()` is executed

Reviewed By: alanz

Differential Revision: D70884595

fbshipit-source-id: df8b7f01e4246db4310b0fadb41bffec4b7814db