Update format of "launch" request
Summary:
# Context
The "launch" and "attach" requests are the two main ways to start debugging sessions in the DAP specification. The payload of both requests are implementation dependant.

For the "launch" request, the server can either start the debuggee process directly and forward stdin/stdout, or, if the client supports it, send a "runInTerminal" reverse-request to the IDE, so that it launches the process inside one of its integrated terminals, etc.

At the moment, "runInTerminal" is the only way in which "launch" works in edb

# Problem
There are a couple of problems in the way we handle the arguments to the "launch" request, which we'll try to address here

  - We are using runInTerminal without checking the client supports it
  - The runInTerminal request allows us to customize several things, like "window title", or env vars, but we don't offer ways for the user to customize all of these

# This diff
We change the format of the "launch" request as follows.

- Before we had a `"launchCommand" field, with an object that needs to be mapped to the arguments of the `runInTerminal" request. Now, we replace it by a "runInTerminal" field that contains exactly an object that matches the arguments of the DAP's `RunInTerminal" spec. We will send this object back (with minor tweaking) to the client when sending the "RunInTerminal" request, which means that whatever the DAP spec supports, we will offer it directly to users.

 - Before, "launchCommand" was part of a "config" field of the request; now instead we move it to the top-level. The idea is to make it more clear that this is "launch with runInTerminal" kind of request. In the future, if we decide to support other modes of launching, we include some other field.

# Caveat

There are still version of the VSCode extension that assume that "launchCommand" is available top-level ; until those are updated and deployed, we need to support them somehow. For that, the "launch" request handler will check the fields of the old "launchCommand" format the the extension may modify, and adapt them to the new format.

Because we have this new provision, we don't need the previous hack where we would accept configuration either under a "config" field or flat, and we can require that the two main fields are always present, simplifying the overall logic.

Reviewed By: alanz

Differential Revision: D70696616

fbshipit-source-id: 92dba2d17b85f4fc8bf4a0489cd29a3943f4fb9c