Make "launch" request use reverse-attach
Summary:
# Context
The "launch" request is the mechanism for an IDE to ask the debugger to start a debuggee process.

The way it currently works, the user is expected to provide the name the new node is expected to have, and the DAP server will try to attach to it, and keep retrying until the timeout.

# Problem
There is a clear race-condition here. For example, if the debuggee node will be running a testcase, then by the time the DAP server attaches, the test is  potentially already running, and past the breakpoint the user wanted to set.

# This diff
We make it so that the "launch" request now uses the "reverse-attach" mechanism implemented earlier. Instead of providing the node-name in advance, the caller now receives in an environment variable a snippet ot Erlang code that they must ensure the debuggee node runs on initialization. This snippet will ensure the debuggee gets attached by edb and paused.

At this point the DAP server enters the configuration phase, and will set all the breakpoints, with the debuggee paused, until the client sends a configurationDone message.

Reviewed By: alanz

Differential Revision: D70795943

fbshipit-source-id: f0367b806c98daa446268c1516394fbe233ae36b