Preserve old value of ERL_AFLAGS
Summary:
# Context
When using the "launch" request to start a new debuggee node, the DAP server sends back "runInTerminal" request to the IDE, that includes the command to be run by the IDE, but also environment variable overrides.

We let the user configure the environment overrides the want in the "launch" request configuration, but the DAP server will also modify the ERL_AFLAGS variable.

# Problem
While we take into account the case when the user asked to set a specific value in ERL_AFLAGS, and ensure we prepend what we need to that; if the user has an ERL_AFLAGS already set in their environment, we are going to trash it and overwrite with what the DAP server needs

# This diff
If the user didn't specify an override for ERL_AFLAGS, we now check the environment and, if this is set, we amend that value.

Reviewed By: alanz

Differential Revision: D70898562

fbshipit-source-id: fb51f7eac1fa8bf16ed61dcdb48ac24fabdd3d09