Reduce boilerplate in tests
Summary:
# Context
Most tests for the DAP server need to start a session, including a debuggee node. The way we are doing it, each test-case needs to start a debuggee node, and pass it to a function that starts the DAP server and client and attach to that node.

# Problem
We now have support functions to start a session using "launch", and these functions create also the debuggee under the hood. This let us simplify a bit the setup of testcases, so that each test focuses on what is being teted.

Reviewed By: alanz

Differential Revision: D70972698

fbshipit-source-id: 8c7bcb96381f49477794004f66b6086c782e9c07