Allow to wait for reverse-request in test client
Summary:
# Context
In tests for the DAP server, we use a dummy client to send a receive DAP requests

# Problem
At the moment, the client offers a function to wait for asynchronous events coming from the server. However, the server can also send async "reverse" requests and currently the client immediately cknowledges them, and we can't inspect their value received on tests. This limits what we can test.

# This diff
Extend the client so we now need to explicitly wait for reverse-requests and respond to them.

Reviewed By: Balajiganapathi

Differential Revision: D70775075

fbshipit-source-id: 37166db334225a4da3e089aa2d7785fb116b3b80