Add an edb_dap_request:abort() function to short-circuit request handlers
Summary:
# Context
The DAP server is essentially a `gen_server` that serves mostly requests coming from the IDE. For each type of request, we define `handle()` function, that returns a `reaction()` term, telling the `gen_server` what to do next: update the state, send a response to the IDE, send some events to the IDE, etc.

# Problem
Because the handler needs to return an "error reaction" when there are user errors, we are forced to use a style where every check opens case, and the happy path keeps getting indented more and more to the right. This makes the main logic harder to follow.

# This diff
We introduce a `edb_dap_request:abort()` function that we can use to short-circuit the execution of the handler, returning an error. With this, we "linearize" a couple of handlers, make the logic hopefully more clear.

Reviewed By: alanz

Differential Revision: D70782755

fbshipit-source-id: 51522b81224a5ce028c0648cdfcb9eff59cce3d9