Make "disconnect" terminate the debuggee if it was "launched"
Summary:
# Context
In the DAP specification, there are two ways to connect to a "debuggee", using the "attach" or the "launch" requests. The former is expected to debug a process that is already running, the latter is expected to launch a new process to be debugged.

For ending a debugging session, there is a single "disconnect" request. The [spec](https://microsoft.github.io/debug-adapter-protocol/specification#Requests_Disconnect) says

>  [T]he debug adapter must terminate the debuggee if it was started with the launch request. If an attach request was used to connect to the debuggee, then the debug adapter must not terminate the debuggee.

From VSCode side, this is reflected by the fact that when "launching", the debugger bar shows a "stop" button (red square), while when "attaching" it shows a "disconnect" button (two red cables or something like that)

# Problem
We are currently not following the specification and we never terminate the debuggee, regardless of how it was started

# This diff
We do two things:

1. We extend the DAP server state to track information on whether the debuggee was "launched" or "attached"
2. When handling the "disconnect" request, we decide whether to kill or not the debuggee based on this.

How do we kill the debuggee?

  - First we try with `erlang:halt(0)` (so we don't produce a crashdump) and if this doesn't seem to finish, we `kill -9` the process (whose OS PID we saved when first connected to the node)
 - When using using "launch", we run the debuggee via the IDE's terminal, and we may get the PID of the shell. If we get it, we kill that PID as well, so that the debugging session also closes from the IDE's point of view.

Reviewed By: robertoaloi

Differential Revision: D70977834

fbshipit-source-id: eac7b4a4cdf233d131d7db699652d441fd8ba701