Support the terminateDebuggee capability
Summary:
# Context
When the DAP server declares that it supports the `terminateDebuggee` capability, then the client can include a `terminateDebuggee` field in the `disconnect` request, to override the default behaviour of terminating the debuggee  if-and-only-if it was launched (instead of attached).

From VSCode side, this means that, e.g., one gets the "stop" button on the debugger bar when the debuggee was "launched", but when pressing "shift" with the mouse on top of the button, it transforms into a "disconnect" button.

# Problem
We are currently not supporting this capability.

# This diff
Add support for it. In the process, we add validation for the "initialize" and "disconnect" requests, so we don't crash if the client sends  bad input.

Reviewed By: robertoaloi

Differential Revision: D70981115

fbshipit-source-id: 5572f8f56aef26c38c5fb65670a500f08968164d