Make stack-frames building function parametric
Summary:
# Context
The stack-frames that we get from the VM via the erl_debugger module
sometime contain frames that we want to hide from the user. This
is typically the case for the frames that are generated while handling
a brekapoint.

This is done in `edb_server_inspect:format_stack_frames/2`, that
skips those frames and build `edb:stack_frame()` terms.

# Problem
For the sake of processing stepping, we want to generate summaries of stack frames that contain just the return addresses. This needs to be subject to the same rules for "ignoring frames" as described above, and we don't want to duplicate this logic.

# This diff
We make it so that `format_stack_frames()` now takes a formatter as argument, so that we will be able to use it to return different types of frames. This happens later in the stack

Reviewed By: Balajiganapathi

Differential Revision: D71334854

fbshipit-source-id: 1b9e7bec8d3e5ef0fbf2453fe0e1f77ba38159d6