Decompose function for formatting stack frames
Summary:
# Context
When returning stack frames to the user, we need to:

1) Select which frames are relevant to the user. E.g. the frames that happened while handling the breakpoint can be ignored, and the bottom one with the terminator is also not interesting
2) Convert from an `erl_debugger:stack_frame()` to the type we care about

# Problem
The current logic is all mixed up and makes the intention harder to understand.

# This diff
We can make it easier by having simpler functions to deal with each aspect and let the caller reassemble. Also, in future diffs, this will give us the option of filtering out fewer frames.

Reviewed By: Balajiganapathi

Differential Revision: D71410091

fbshipit-source-id: cf5be1e6abecba2256b9f5ead1ae8303bca07a2b