Pre-emptively suspend process when process in step breakpoint
Summary:
# Context
The goal of the stack is to make step-over, etc, work in the context of recursion. Here the problem is that it is not enough to set a special breakpoint on the functions we expect to land while stepping, we also need to remember what the call-stack looked like when we started stepping, to know if we are in the same instance of the function or a new recursive call.

# Problem
The above means that when processing a "step" breakpoint, we will need to find out details about the call-stack of the process hitting the breakpoint. However, the erl_debugger API asks the process to be suspended in order to give us stack-information.

The problem is that:

  1. We suspend all processes after we decide to accept the breakpoint,
  2. We need to resume the process as many times as we suspended it. However, if we do it here, we incur the overhead of pausing the timers, to unpause them immediately, to pause them again.

# This diff
So we establish the following protocol:

when processing a breakpoint, if it turns out to be a `step`, we assume the process has been suspended. That means that we need to either avoid suspending it again, or we need to resume it afterwards to balance the count.

We do the latter as it is simpler. Notice that once the process was suspended the first time, suspending it again and then resuming it once are cheap, as they only increase/decrease a counter.

Reviewed By: TheGeorge

Differential Revision: D71412073

fbshipit-source-id: f19ebfe010af37008083d42e096de16eb31415b6