Simplify stack-frame handling
Summary:
# Context
The "raw" stack frames for a process hitting a breakpoint contain extra frames that are implementation details and we want to hide from the user in general.

In D71410091 we simplified some of the functions we had for dealing with this, in particular we a function to just remove those extra frames

# Problem
We still have some ad-hoc functions to check the the "frame-id" of the first "user frame" in a raw stack trace. This is now redundant as we can just filter out the frames we don't care about and get the top

# This diff
Simplify the code by removing that ad-hoc function and relying on the recently introduced `user_frames_only()`

Reviewed By: Balajiganapathi

Differential Revision: D71494622

fbshipit-source-id: 68b44ee35b8584daacce59202c226eb1daefddf2