Use raw stack-frames when setting stepping breakpoints
Summary:
# Context
To implement stepping operations, we need to set breakpoints in the functions along the current call-stack.

For this, we currently reuse the same functions we use to return stack-frames to the user. This means that we take the raw stack-frames returned by the `erl_debugger` module, and "format" them, mostly removing information we don't expose to users

# Problem
There's information in the raw-frames that we want to use for the book-keeping of stepping breakpoints, so it's not available in the current interface

# This diff
We change it so that we use raw stack-frames when calling the functions to set breakpoints along the call-stack. At the moment this is a no-op change, but we will use the extra information next

Reviewed By: alanz

Differential Revision: D71494621

fbshipit-source-id: d3d3c5b8799aa4c74195d11a57030b94782e43d7