Move logic to decide which frames to intrument on step to correct module
Summary:
# Context
The `edb_server_break` module manages all the breakpoints set on the VM. These can be "explicit" breakpoints, set by the user, or internal breakpoints set for implementing the stepping primitives.

When setting breakpoints for the latter, we need to chose the stack-frames whose function need to be instrumented with breakpoints.

# Problem
Ideally, the `edb_server_break` module is the one that would decide which frames to instrument. Instead, this is responsibility of the caller.

The reason for this is historical: when this code was first written, we reused logic to obtain stack-frames that was already in `edb_server` and we didn't want to export the relevant functions and introduce a dependency.

# This diff
Since we have previously refactored things, now we can have `edb_server_break` depend on `edb_server_stack_frames` for all its stack-frames need, thus, clean-up its interface

Reviewed By: alanz

Differential Revision: D71549089

fbshipit-source-id: 6495211282e51aa3dd83a65847a7abd07faf2f1f