Use dedicate module for dealing with stack-frames
Summary:
# Context
At the moment we have a module called `edb_server_inspect`, which is a form of "utility" module. Most exported functions there have to do with processing stack-frames.

In `edb_server` (the module with the gen_server implementation) we also have a couple of internal functions for getting the stack-frames for a process, and do error-handling.

# Problem
It is not clear what "inspect" is about, so the module is confusing. Moreover, it would be make for better API if we were to retrieve stack-frames from the module handling breakpoints/stepping (to be done next in the stack), but it doesn't make sense to export the functions we already have for this from `edb_server`

# This diff
We rename `edb_server_inspect` into `edb_server_stack_frames`, and now it handles all things stack-frames.

We move the function that was finding the source for a module to `edb_code`, which is another kind of "utillity" module, but at least the name sounds more appropriate, for now

Reviewed By: alanz

Differential Revision: D71540965

fbshipit-source-id: 088c9c5f7aaa4123eb5d8d1bc33332cb3ead9a4b