Workaround bug in otp parser annotation
Summary:
# Context
To implement stepping, roughly speaking, we set a breakpoint on the lines of the current function. For this, given a line number, we find in the AST in the debug chunk of the beam for the function that contains it, and then find all the lines in that ast for that function.

# Problem
The latest version of the patched otp toolchain seems to bring in an upstream bug. Essentially, for a term `[a, b, c]` (on line 113), the ast looks like this:

```
     {cons, {113,9}, {atom,{113,10},a},
     {cons,  {113,13}, {atom,{113,13},b},
     {cons,{113,16},{atom,{113,16},c},
     {nil,0}}}},
```

Notice that each `cons` comes with the right line-col, but the final `nil` has 0 as annotation.

We then consider line 0 as a line in the function (and actually, all lines from 0 onwards) and when we try to set a breakpoint on line 0, the BIF fails.

Unfortunately, our test suite doesn't contain a case where we step on a function with a list term, so we missed this when we recently updated the toolchain

# This diff
We need to follow up this upstream, but until then, let's ignore any 0 we may find

Reviewed By: robertoaloi

Differential Revision: D71663424

fbshipit-source-id: 5995bd278e62550d6ede216ab44be66b13cac58f