Add dummy request handler for stepIn request
Summary:
# Context
We are in the process of implementing the support for stepping-in a function call. Because the functionality is not ready, there was no request handler in the DAP server for the "stepIn" request.

# Problem
There's no optional capability for stepping-in in the DAP specification so VSCode assumes that step-in works, so when the user tries to step-in, they get a generic "Method not found" error, that makes it look like it was an unexpected failure.

# This diff
We add a dummy request handler for "stepIn", and make it return an error that clarifies what's going on.

Reviewed By: alanz

Differential Revision: D71674995

fbshipit-source-id: 71fc776d56cbaf1afb5eccfb3fd49659b9576f27