Make stepping work correctly under recursion
Summary:
# Context
We currently have two stepping primitives implemented: step-over and step-out. They work by traversing the call stack and add breakpoints, limited to the process being stepping, on all the functions along the call-stack.

# Problem
This strategy clear breaks when dealing with recursive functions. For example:

  - If we try to "step-over" a recursive call, because we are adding breakpoints on the current function, the process will break during the recursive call, and since we only check that we are expecting that process to break, we will stop. But we will have done a step-in, instead of a step-over

  - Similarly, if we have mutually recursive functions `f` and `g`, and we are stopped in `f` with `g` already on the call stack, if we try to step-out of `f` when we are about to call `g`, then because `g` is in the call stack, we will add breakpoints to `g`, and end up doing a sort of step-in.

# This diff
So to solve this, we not only need to  remember which process we want to stop with those breakpoints, we need to record as well what we expect the stack to look like when this happens.

For this, we leverage the "code address" information that we have in `erl_debugger:stack_frame()`. Essentially, this is the adrress in memory of the next processor instructor that will be executed when that frame resumes. So for all the frames below the top frame, we expect that value to match. For the top frame, we don't know what the value will be so we use a wildcard.

Although not strictly necessary for this diff, we use the MFA of the frame as wildcard. This will be convenient when adding support for step-in, as in that case, we won't know the expected address of  the two topmost frames, so using MFA on the second one wil be a good sanity check

Reviewed By: alanz

Differential Revision: D71412010

fbshipit-source-id: 67952ec50be042b3b869949991ee63be4511f6a0