Simplify logic to set breakpoints for stepping
Summary:
# Context
To implement stepping we set breakpoints on all functions on the stack-frame. One problem we have is what happens if, say, the otp code is not compiled with breakpoints support: if fail to set breakpoints and say nothing, the user will get confused when stepping doesn;t stop; but if we reject to step, then in many cases stepping will just not work ever.

The compromise is to enforce that at least the "top frame" has debugging support.

# Problem
To implement this logic we have two very similar functions, and the names don't help clarifying what's the difference between them.

# This diff
Merge them into one function, with an arg that explains better what the difference is is

Reviewed By: alanz

Differential Revision: D71681130

fbshipit-source-id: acb2c265bfdd2f360fe2e5dee19ec5d7997783c2