Unify logic for compiling test modules in tests
Summary:
# Context
In most edb tests, we want to compile some test modules at test runtime, and then make the debuggee run code in that module.

# Problem
For historical reasons, we have two separate logics to compile and load modules, that evolved independently: that in the DAP tests and that in the core api tests.

The one for DAP is a bit more advanced, as it can handle modules defined as resource files, or defined inline in the testcase using their contents.  So it would be good to merge them, so we can eventually take advantage of the latter in core api tests.

# This diff
We expose a `edb_test_support:compile_module/1` that is now used by both families of tests. We need to extend it to handle scenarios that are needed in the core api tests: overriding the compiler flags, and loading the module afterwards.

Reviewed By: alanz

Differential Revision: D71720712

fbshipit-source-id: 16bc559343340b431bfd46485e446d2790cb4eff