Use erl_syntax instead of manipulating parse tree
Summary:
# Context
For stepping, we use the AST of the module to find the lines where we need to add breakpoints, etc. This logic is in the `edb_server_code` module

# Problem
At the moment, we are pattern-matching on the `erl_parse` types directly. These types may change over time, it is much more robust to use `erl_syntax` instead.

# This diff
Use `erl_syntax` instead instead of pattern-matching on the `erl_parse` terms

Reviewed By: RobinMorisset

Differential Revision: D71720711

fbshipit-source-id: a01cf691ae340d13527e2a7d51fb1e18b64488a7