Separate reading forms from finding lines for function
Summary:
# Context
Each time we want to step-over, we take the line at which the process is stopped, and set breakpoints on all the lines for the function that contains that line.

For this, we use the AST of the module that is stored in the beam when compiling with `+debug_info`

# Problem
We are currently reading the AST from the beam file every time we call the function to find the lines of a function. As we add more analyses of the AST (e.g., for step-in), we will want to avoid re-reading the AST over and over again. Eventually, we may even want to read it, index it, and save the index in some cache.

# This diff
For now, we decouple reading the AST from finding the lines. That way, the caller is free to reuse the same AST between analyses

Reviewed By: alanz

Differential Revision: D71720715

fbshipit-source-id: 42354ea472d4d36f2f3095e76e10faf9deda160a