Make it easier to set step-breakpoints on a function
Summary:
# Context
To implement stepping, we set breakpoints on all the lines of functions along the call-stack, and keep some book-keeping to know which processes they correspond to.

At the moment we support step-over and step-out and in both cases, it is enough to know a source line (for step-over, where we are currently standing, for the call-stack, the location of the caller). So the support functions we have take a line-number and find the function containing it and from there get the lines.

# Problem
We want to implement step-in, and in that case, we will know the MFA we want to step-into, but not any of their line numbers. So we don't currently have an easy way to set breakpoints on those

# This diff
We refactor things a bit so that we can in the future set breakpoints on a given function. This means:

  - In `edb_server_break` and an auxiliary function that sets brekapoints on a function, given its AST
  - In `edb_server_code` we decouple further the functions we were doing to perform analyses

In doing this, we clean up a bit the logic to detect if we didn't manage to set any breakpoint, avoiding having to compare maps for this

Reviewed By: alanz

Differential Revision: D71756174

fbshipit-source-id: e05a7045af07316ca76523d18369b456e2b2f18b