Add simple analysis to identify call targets
Summary:
# Context
We want to implement the step-in functionality. The strategy is: check if we are standing on a call, figure out the target of the call, and set breakpoints there.

# Problem
We need to have some kind of analysis that gives us the call-target of a call, based on the current line

# This diff
We introduce a `edb_server_code:get_call_targets/2` function that, given a line, checks if it is a call and returns, if supported, the target of the call.

For now, we only support named, static calls:

Reviewed By: alanz

Differential Revision: D71720713

fbshipit-source-id: 094b4dd64f1088dccce83531b494cb81ff411192