Handle race-condition in "terminate" request
Summary:
# Context
The DAP client sends a "disconnect" request to end the debugging session. This can happen as a result of the user clicking the "stop" button, but also after the DAP server indicating that the debuggee is down.

# Problem
We currently try to stop the debugger, which makes sense when this is triggered by the user (e.g on an "attach" session). But if the debuggee has already stopped, then this will raise a `not_attached` error, which will make the request fail.

# This diff
Handle the `not_attached` error, so that we don't abort the request, and don't end up with errors in the logs

Reviewed By: acw224

Differential Revision: D72166843

fbshipit-source-id: 4a6c359a674a6079836b6d9447788c4cdee084fc