Make the process info fields configurable
Summary:
# Context
We use `edb:processes()` to retrieve information about the processes being debugged, and `edb:excluded_processes()` for those not under debugging. At the moment, we return a large and fixed number of fields for each process.

# Problem
There are multiple reasons why we would like the user to control what fields they want to receive, similar to `erlang:process_info/2`. The main two are:

  - When debugging nodes with tens of thousands of processes, it is important to avoid the overhead of computing fields that the caller will not use
  - It gives as a clean path for adding more specific fields or expensive fields that we don't always want to return

# This diff
We change it so that all functions returning function information receive a list with the fields they want to receive.

Reviewed By: TheGeorge

Differential Revision: D72162015

fbshipit-source-id: cf04bb5260512a9ccd6f75ce5ecc34b5764bca71