Rename field in excluded_process_info()
Summary:
# Context
We have functions to return info of processes and "excluded" processes (those not subject to debugging).

They return slightly disjoint fields, so we are using distinct types for them. We want to make the fields returned selectable by the user, so it will be easier if we return terms of the same type in both cases.

# Problem
There is a field called `reason` in the type for excluded-processes, and this name doesn't make sense in a general "process info" type.

# This diff
In order to merge them, we first rename `reason` to `exclusion_reasons`

Reviewed By: TheGeorge

Differential Revision: D72162016

fbshipit-source-id: 86e7c8538842d7eebaf124befe073ec8eb853b4c